move migration policy helper functions to a dedicated file
